Change the optimization level
For some reason, if the optimization level is 1 (i.e., -O or -O1),
we get a segfault in the libudev (only in debian), the problem
is the add_device() call inside the even loop (without that call
everything works).

I'm not sure about the root cause yet, so let's change the
optimization level to allow run it in debian 11.

Signed-off-by: German Maglione <germanm@gmail.com>